copy symbols from loaded files into global namespace
